+++ lib/db/account.dart
@@ -60,7 +60,7 @@ class Account implements Model, TransferTarget {
       icon: IconData(map['icon_code'], fontFamily: map['icon_font']),
       color: Color(map['color']).withOpacity(1),
       archived: map['archived'] == 1,
-      currency: Currency.fromISOCode(map['currency']),
+      currency: Currency.mustFromISOCode(map['currency']),
       order: map['order'],
       balance: map['balance'] ?? 0.0,
     );