+++ lib/providers/account.dart
@@ -2,7 +2,6 @@ import 'dart:collection';
 
 import 'package:fb/db/account.dart';
 import 'package:fb/db/repository.dart';
-import 'package:fb/db/account.dart';
 import 'package:fb/utils/currency.dart';
 import 'package:flutter/material.dart';