+++ lib/providers/transaction.dart
@@ -4,7 +4,6 @@ import 'package:fb/db/repository.dart';
 import 'package:fb/db/transaction.dart';
 import 'package:fb/db/account.dart';
 import 'package:fb/db/category.dart';
-import 'package:fb/db/transaction.dart';
 import 'package:fb/db/transfer_target.dart';
 import 'package:fb/providers/account.dart';
 import 'package:fb/providers/category.dart';